requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
pymongo[srv]
dnspython
pymongo==4.1.1
python-dotenv==0.19.2
openpyxl==3.0.9
xlrd==2.0.1
flask-login==0.5.0
python-dateutil==2.8.2